Capture a screenshot

For the moment, the capture file is in the current working directory and named screenshot.png. This will change, soon.

chrome_shot(url, width = NULL, height = NULL)

Arguments

url
URL to read from

width, height
screen size to emulate

Value

magick

Details

A magick image object is returned.

Note

this is a quick version of the function and will overwrite screenshot.png if it exists in CWD

Examples

chrome_shot("https://www.r-project.org/logo/Rlogo.svg")

#> Warning: error in running command

#> Error in magick_image_readpath(path, density, depth, strip, defines): R: unable to open image `screenshot.png': No such file or directory @ error/blob.c/OpenBlob/2874